+++ src/App.jsx
@@ -1,13 +1,13 @@
 import './styles/App.css';
 import NavbarMobile from './components/NavbarMobile';
-import { Routes, Route, useNavigate } from 'react-router-dom';
+import { Routes, Route } from 'react-router-dom';
 import About from './pages/About';
 import Projects from './pages/Projects';
 import Education from './pages/Education';
 import Interest from './pages/Interest';
 import Journey from './pages/Journey';
 import Skills from './pages/Skills';
-import { useEffect, useState } from 'react';
+import PageNotFound from './pages/PageNotFound';
 
 const navLinks = [
   'About', //don't change the order of these links
@@ -34,30 +34,5 @@ function App() {
     </>
   );
 }
-
-const PageNotFound = () => {
-  const [count, setCount] = useState(3); // [0, function
-  const navigate = useNavigate();
-  useEffect(() => {
-    setTimeout(() => {
-      console.log('Redirecting to About page...');
-      navigate('/About');
-    }, 3000);
-  }, []);
-
-  useEffect(() => {
-    setTimeout(() => {
-      setCount(count - 1);
-    }, 1000);
-  }, [count]);
-
-  return (
-    <div>
-      <h1>Error 404</h1>
-      <h1>Page Not Found</h1>
-      <h1>Redirecting to About page... {count}</h1>
-    </div>
-  );
-};
 
 export default App;
